Add ST API wrappers to bridge the low-level callbacks to async functions
